Refactor add-action-failure
